vidation text, show select, remove newPass_new
